requirements.txt
streamlit
streamlit-webrtc
opencv-python-headless
Pillow
numpy